Module Slas modifications to use inheritance
