+++ LibreNMS/OS/Shared/Cisco.php
@@ -28,6 +28,7 @@ namespace LibreNMS\OS\Shared;
 use App\Models\Device;
 use App\Models\Mempool;
 use App\Models\PortsNac;
+use App\Models\Sla;
 use Illuminate\Support\Arr;
 use LibreNMS\Device\Processor;
 use LibreNMS\Interfaces\Discovery\MempoolsDiscovery;
@@ -302,6 +303,77 @@ class Cisco extends OS implements OSDiscovery, ProcessorDiscovery, MempoolsDisco
         return $processors;
     }
 
+    public function discoverSlas()
+    {
+        $device = $this->getDeviceArray();
+
+        $slas = collect();
+        $data = snmp_walk($device, 'ciscoRttMonMIB.ciscoRttMonObjects.rttMonCtrl', '-Osq', '+CISCO-RTTMON-MIB');
+
+        // Index the MIB information
+        $sla_table = [];
+        foreach (explode("\n", $data) as $index) {
+            $key_val = explode(' ', $index, 2);
+            if (count($key_val) != 2) {
+                $key_val[] = '';
+            }
+
+            $key = $key_val[0];
+            $value = $key_val[1];
+
+            $prop_id = explode('.', $key);
+            if ((count($prop_id) != 2) || ! ctype_digit($prop_id[1])) {
+                continue;
+            }
+
+            $property = $prop_id[0];
+            $id = intval($prop_id[1]);
+
+            $sla_table[$id][$property] = trim($value);
+        }
+
+        foreach ($sla_table as $sla_nr => $sla_config) {
+            $data = [
+                'device_id' => $device['device_id'],
+                'sla_nr'    => $sla_nr,
+                'owner'     => $sla_config['rttMonCtrlAdminOwner'],
+                'tag'       => $sla_config['rttMonCtrlAdminTag'],
+                'rtt_type'  => $sla_config['rttMonCtrlAdminRttType'],
+                'status'    => ($sla_config['rttMonCtrlAdminStatus'] == 'active') ? 1 : 0,
+                'opstatus'  => ($sla_config['rttMonLatestRttOperSense'] == 'ok') ? 0 : 2,
+                'deleted'   => 0,
+            ];
+
+            // Some fallbacks for when the tag is empty
+            if (! $data['tag']) {
+                switch ($data['rtt_type']) {
+                    case 'http':
+                        $data['tag'] = $sla_config['rttMonEchoAdminURL'];
+                        break;
+
+                    case 'dns':
+                        $data['tag'] = $sla_config['rttMonEchoAdminTargetAddressString'];
+                        break;
+
+                    case 'echo':
+                        $data['tag'] = IP::fromHexString($sla_config['rttMonEchoAdminTargetAddress'], true);
+                        break;
+
+                    case 'jitter':
+                        if ($sla_config['rttMonEchoAdminCodecType'] != 'notApplicable') {
+                            $codec_info = ' (' . $sla_config['rttMonEchoAdminCodecType'] . ' @ ' . preg_replace('/milliseconds/', 'ms', $sla_config['rttMonEchoAdminCodecInterval']) . ')';
+                        } else {
+                            $codec_info = '';
+                        }
+                        $data['tag'] = IP::fromHexString($sla_config['rttMonEchoAdminTargetAddress'], true) . ':' . $sla_config['rttMonEchoAdminTargetPort'] . $codec_info;
+                        break;
+                }//end switch
+            }//end if
+            $slas->push($data);
+        }
+        return $slas;
+    }
+
     public function pollNac()
     {
         $nac = collect();
@@ -344,6 +416,116 @@ class Cisco extends OS implements OSDiscovery, ProcessorDiscovery, MempoolsDisco
         }
 
         return $nac;
+    }
+
+    public function pollSlas($slas)
+    {
+        $device = $this->getDeviceArray();
+
+        // Go get some data from the device.
+        $rttMonLatestRttOperTable = snmpwalk_array_num($device, '.1.3.6.1.4.1.9.9.42.1.2.10.1', 1);
+        $rttMonLatestOper = snmpwalk_array_num($device, '.1.3.6.1.4.1.9.9.42.1.5', 1);
+
+        $uptime = snmp_get($device, 'sysUpTime.0', '-Otv');
+        $time_offset = (time() - intval($uptime) / 100);
+
+        echo "\nSLAS COUNT ==> " . count($slas) . "\n";
+        foreach ($slas as $sla) {
+            $sla_id = $sla['sla_id'];
+            $sla_nr = $sla['sla_nr'];
+            $rtt_type = $sla['rtt_type'];
+
+            // Lets process each SLA
+            $unixtime = intval(($rttMonLatestRttOperTable['1.3.6.1.4.1.9.9.42.1.2.10.1.5'][$sla_nr] / 100 + $time_offset));
+            $time = strftime('%Y-%m-%d %H:%M:%S', $unixtime);
+            $update = [];
+
+            // Use Nagios Status codes.
+            $opstatus = $rttMonLatestRttOperTable['1.3.6.1.4.1.9.9.42.1.2.10.1.2'][$sla_nr];
+            if ($opstatus == 1) {
+                $opstatus = 0;        // 0=Good
+            } else {
+                $opstatus = 2;        // 2=Critical
+            }
+
+            // Populating the update array means we need to update the DB.
+            if ($opstatus != $sla['opstatus']) {
+                $update['opstatus'] = $opstatus;
+            }
+
+            $rtt = $rttMonLatestRttOperTable['1.3.6.1.4.1.9.9.42.1.2.10.1.1'][$sla_nr];
+            echo 'SLA ' . $sla_nr . ': ' . $rtt_type . ' ' . $sla['owner'] . ' ' . $sla['tag'] . '... ' . $rtt . 'ms at ' . $time . '\n';
+
+            $fields = [
+                'rtt' => $rtt,
+            ];
+
+            // Let's gather some per-type fields.
+            switch ($rtt_type) {
+                case 'jitter':
+                    $jitter = [
+                        'PacketLossSD' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.26'][$sla_nr],
+                        'PacketLossDS' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.27'][$sla_nr],
+                        'PacketOutOfSequence' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.28'][$sla_nr],
+                        'PacketMIA' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.29'][$sla_nr],
+                        'PacketLateArrival' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.30'][$sla_nr],
+                        'MOS' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.42'][$sla_nr] / 100,
+                        'ICPIF' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.43'][$sla_nr],
+                        'OWAvgSD' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.49'][$sla_nr],
+                        'OWAvgDS' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.50'][$sla_nr],
+                        'AvgSDJ' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.47'][$sla_nr],
+                        'AvgDSJ' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.2.1.48'][$sla_nr],
+                    ];
+                    $rrd_name = ['sla', $sla_nr, $rtt_type];
+                    $rrd_def = RrdDefinition::make()
+                        ->addDataset('PacketLossSD', 'GAUGE', 0)
+                        ->addDataset('PacketLossDS', 'GAUGE', 0)
+                        ->addDataset('PacketOutOfSequence', 'GAUGE', 0)
+                        ->addDataset('PacketMIA', 'GAUGE', 0)
+                        ->addDataset('PacketLateArrival', 'GAUGE', 0)
+                        ->addDataset('MOS', 'GAUGE', 0)
+                        ->addDataset('ICPIF', 'GAUGE', 0)
+                        ->addDataset('OWAvgSD', 'GAUGE', 0)
+                        ->addDataset('OWAvgDS', 'GAUGE', 0)
+                        ->addDataset('AvgSDJ', 'GAUGE', 0)
+                        ->addDataset('AvgDSJ', 'GAUGE', 0);
+                    $tags = compact('rrd_name', 'rrd_def', 'sla_nr', 'rtt_type');
+                    data_update($device, 'sla', $tags, $jitter);
+                    $fields = array_merge($fields, $jitter);
+                    break;
+                case 'icmpjitter':
+                    $icmpjitter = [
+                        'PacketLoss' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.26'][$sla_nr],
+                        'PacketOosSD' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.28'][$sla_nr],
+                        'PacketOosDS' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.29'][$sla_nr],
+                        'PacketLateArrival' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.32'][$sla_nr],
+                        'JitterAvgSD' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.45'][$sla_nr],
+                        'JitterAvgDS' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.46'][$sla_nr],
+                        'LatencyOWAvgSD' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.47'][$sla_nr],
+                        'LatencyOWAvgDS' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.48'][$sla_nr],
+                        'JitterIAJOut' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.49'][$sla_nr],
+                        'JitterIAJIn' => $rttMonLatestOper['1.3.6.1.4.1.9.9.42.1.5.4.1.50'][$sla_nr],
+                    ];
+                    $rrd_name = ['sla', $sla_nr, $rtt_type];
+                    $rrd_def = RrdDefinition::make()
+                        ->addDataset('PacketLoss', 'GAUGE', 0)
+                        ->addDataset('PacketOosSD', 'GAUGE', 0)
+                        ->addDataset('PacketOosDS', 'GAUGE', 0)
+                        ->addDataset('PacketLateArrival', 'GAUGE', 0)
+                        ->addDataset('JitterAvgSD', 'GAUGE', 0)
+                        ->addDataset('JitterAvgDS', 'GAUGE', 0)
+                        ->addDataset('LatencyOWAvgSD', 'GAUGE', 0)
+                        ->addDataset('LatencyOWAvgDS', 'GAUGE', 0)
+                        ->addDataset('JitterIAJOut', 'GAUGE', 0)
+                        ->addDataset('JitterIAJIn', 'GAUGE', 0);
+                    $tags = compact('rrd_name', 'rrd_def', 'sla_nr', 'rtt_type');
+                    data_update($device, 'sla', $tags, $icmpjitter);
+                    $fields = array_merge($fields, $icmpjitter);
+                    break;
+            }
+
+            return array($fields, $update);
+        }
     }
 
     protected function getMainSerial()
